requirements.txt
aiohttp
asyncio
discord.py[speed]
aiofiles
pyyaml
asyncpg